README update ARK dynamic fees implementation ARK packaging improvement
