+++ dposlib/network/prs.net
@@ -1,7 +1,7 @@
 {
 	"timeout": 5,
 	"begintime": [2018, 2, 1, 0],
-	"familly": "ark",
+	"familly": "ark.v1",
 	"maxlimit": 50,
 	"maxvotepertx": 1,
 	"maxvote": 1,
