+++ dposlib/network/ripa.net
@@ -1,7 +1,7 @@
 {
 	"timeout": 5,
 	"begintime": [2017, 3, 21, 13],
-	"familly": "ark",
+	"familly": "ark.v1",
 	"maxlimit": 25,
 	"maxvotepertx": 1,
 	"maxvote": 1,
